cl-svg-polygon now supports holes in paths!
if an svg object has holes in it, it will return under a separate key
from :point-data under :holes. :holes is a *list* of holes, each hole a
*vector* of points (list x y).

this can be shoved into a lib like glu-tessellate
(http://github.com/orthecreedence/glu-tessellate) to triangulate the
points (w/ holes) and display in opengl as triangles.

the *only* path operation that throws an error now is arcs. hopefully
nobody uses them. i sure dont...